fixed some naming inside Angular 2-5 lib, removed unused code
