+++ Mezzurite.Angular2_5/src/angular-performance.module.ts
@@ -1,10 +1,7 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
+﻿import { NgModule, ModuleWithProviders } from "@angular/core";
+import { MezzuriteDirective } from './angular-performance.directive';
+import { RoutingService } from "./angular-routing.service";
 
-import { NgModule, ModuleWithProviders } from "@angular/core";
-import { PerformanceTimingService } from "@microsoft/mezzurite-core";
-import { MezzuriteDirective } from './angular-performance-directive';
-import { RoutingService } from "./routing.service";
 
 @NgModule({
     declarations: [MezzuriteDirective],
@@ -19,7 +16,6 @@ export class AngularPerfModule {
         return {
             ngModule: AngularPerfModule,
             providers: [
-                { provide: PerformanceTimingService, useFactory: createPerfTimingService },
                 RoutingService
                 // Add new services here.
             ]
@@ -27,12 +23,6 @@ export class AngularPerfModule {
     }
 }
 
-export function createPerfTimingService() {
-    return new PerformanceTimingService();
-}
-
 export {
-    PerformanceTimingService,
     RoutingService
-    // Add new services here.
 };
