--- Mezzurite.Angular2_5/src/performance-utils-angular.service.ts
@@ -1,16 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-import { MezzuriteUtils } from "@microsoft/mezzurite-core";
-import {environment} from "./environment";
-
-/**
- * Extension of Mezzurite Utilities that gets package specific information
- */
-export class MezzuriteAngularUtils extends MezzuriteUtils{
-    static createMezzuriteObject(obj: any){
-        super.createMezzuriteObject(obj);
-        (<any>window).mezzurite.packageVersion = environment.version;
-        (<any>window).mezzurite.packageName = environment.name;
-    }
-}
